genai/requirements.txt
streamlit>=1.20
sqlalchemy>=1.4
requests>=2.28
chromadb>=0.3.26
sentence-transformers>=2.2.2
python-dotenv>=1.0
requests
beautifulsoup4
readability-lxml
python-dateutil
tldextract
numpy
streamlit
sqlalchemy
requests
sentence-transformers
numpy
python-dotenv
streamlit
sqlalchemy
requests
sentence-transformers
numpy
cohere

